refactor: modularize cli utilities
