+++ lib/src/cli/cli.dart
@@ -1,0 +1,43 @@
+import 'package:path/path.dart' as p;
+import 'package:universal_io/io.dart';
+
+part 'flutter_cli.dart';
+
+/// Abstraction for running commands via command-line.
+class _Cmd {
+  /// Runs the specified [cmd] with the provided [args].
+  static Future<ProcessResult> run(
+    String cmd,
+    List<String> args, {
+    bool throwOnError = true,
+    String? workingDirectory,
+  }) async {
+    final result = await Process.run(cmd, args,
+        workingDirectory: workingDirectory, runInShell: true);
+
+    if (throwOnError) {
+      _throwIfProcessFailed(result, cmd, args);
+    }
+    return result;
+  }
+
+  static void _throwIfProcessFailed(
+    ProcessResult pr,
+    String process,
+    List<String> args,
+  ) {
+    if (pr.exitCode != 0) {
+      final values = {
+        'Standard out': pr.stdout.toString().trim(),
+        'Standard error': pr.stderr.toString().trim()
+      }..removeWhere((k, v) => v.isEmpty);
+
+      var message = 'Unknown error';
+      if (values.isNotEmpty) {
+        message = values.entries.map((e) => '${e.key}\n${e.value}').join('\n');
+      }
+
+      throw ProcessException(process, args, message, pr.exitCode);
+    }
+  }
+}
