+++ lib/src/cli/flutter_cli.dart
@@ -1,5 +1,4 @@
-import 'package:path/path.dart' as p;
-import 'package:universal_io/io.dart';
+part of 'cli.dart';
 
 /// Thrown when `flutter packages get` or `flutter pub get`
 /// is exectuted without a pubspec.yaml
@@ -79,45 +78,6 @@ class Flutter {
     String cwd = '.',
   }) {
     return Directory(cwd).listSync(recursive: true).where(where).map(run);
-  }
-}
-
-/// Abstraction for running commands via command-line.
-class _Cmd {
-  /// Runs the specified [cmd] with the provided [args].
-  static Future<ProcessResult> run(
-    String cmd,
-    List<String> args, {
-    bool throwOnError = true,
-    String? workingDirectory,
-  }) async {
-    final result = await Process.run(cmd, args,
-        workingDirectory: workingDirectory, runInShell: true);
-
-    if (throwOnError) {
-      _throwIfProcessFailed(result, cmd, args);
-    }
-    return result;
-  }
-
-  static void _throwIfProcessFailed(
-    ProcessResult pr,
-    String process,
-    List<String> args,
-  ) {
-    if (pr.exitCode != 0) {
-      final values = {
-        'Standard out': pr.stdout.toString().trim(),
-        'Standard error': pr.stderr.toString().trim()
-      }..removeWhere((k, v) => v.isEmpty);
-
-      var message = 'Unknown error';
-      if (values.isNotEmpty) {
-        message = values.entries.map((e) => '${e.key}\n${e.value}').join('\n');
-      }
-
-      throw ProcessException(process, args, message, pr.exitCode);
-    }
   }
 }
